==> 1-Lohn.qmd <==
---
bibliography: references.bib
---

# Notes

Scratch Notes for the Seminar Paper

## Sonneberg

Sonneberg: - Germany's first AfD district administrator (Landrat) -
Robert Sesselmann - prevails in run-off election against conservative
CDU - 52.8% to 47.2%

shocking to political establishment, international coverage

Sonneberg also: district with the highest proportion of minimum wage
earners in DE - 2022 increase in minimum wage to 12 euros - 44% of
Sonneberg residents benefited from this - the highest in Germany

Research Question: What is the connection between minimum wage and
populism?

additional about Sonneberg:

-   thüringenweit höchste Industriedichte
-   schon damals viele Arbeiter, die erst rot und dann NS gewählt haben
    (Wiki)
-   Thüringen erste NSDAP Regierungsbeteiligung
    -   [Spiegel](https://www.spiegel.de/geschichte/nsdap-1930-probelauf-der-diktatur-a-e9e83be9-0002-0001-0000-000207602126)
    -   [MDR](https://www.mdr.de/geschichte/ns-zeit/politik-gesellschaft/mustergau-thueringen-gleichschaltung-100.html)
-   auch genannt SPielzeugstadt
-   [Wahlergebnisse
    hsitorisch](https://www.db-thueringen.de/servlets/MCRFileNodeServlet/dbt_derivate_00023432/wahlenundabstimmungsergebnissse_1920-1995.pdf)
    S.90
    -   Stadt mit meisten NSDAP Stimmen 1929

## Literature

### Overview Papers Economic Drivers

@fischerEconomicDeterminantsPopulism2023 (DIW)

-   trade exposure (globalization)
-   financial crises
    -   austerity measures
    -   unfair perception of bailouts
-   MIgration
    -   economic fears (must not be rational)
    -   sudden influxes of many = more right wing
-   inequality
    -   perception of unfairness
    -   low intergenerail mobility

another overview by @gurievPoliticalEconomyPopulism2022

-   trade
    -   exposure to low wage industrial competition
    -   especially in rural regions (cities could protect from impacts
        with service industry)
-   economic crisis
    -   austerity
    -   redistribution
-   cultural backlash
-   immigration
    -   shock immigration of 2015 syrian refugee crisis
    -   higher perception than reality
-   also emigration (esp. eastern europe)
    -   good ground for populist vote of the remaining

### Economic Drivers

Literature on economic drivers presented above, but on german / european district level

@gabrielPoliticalCostsAusterity2023

regional Database on political costs of austerity

-   fiscal austeritiy =\>
    -   higher share for extreme parties
    -   lower vote turnout
    -   1% reduction =\> 3% increase vote
-   austerity driven recession =\> even more
    -   lower trust in government
    -   special of just recession

@dippelEffectTradeWorkers2022 : Trade in Germany Populist

-   case study germany Effect of Trade on Workers and Votes
-   low skilled manufacturing workers = votes responsive to trade
    exposure
-   level of Landkreise in Germany
-   have interesting dataset on trade exposure in the counties

@alesinaIntergenerationalMobilityPreferences2018

-   lower intergenerational mobility
-   higher rpreference for redistributional policy
-   differences in perception

### Sociodemographic Driver Populism

@pickelWahlAfDFrustration2019: Sozialstrukturelle Analyse

-   self identify class: lower class & working class & middle class
    -   real: mostly workers and Facharbeiter
    -   also some lower state servants (Beamte)
-   no age / income differences to general population

@hoevermannUmfragehochAFD2023: Sociodemografic Factors and Work

-   more temporary workers (befristet)
-   bad working conditions = lower esteem of work
-   uncontempt with current wage
-   lower share of workers council (Betriebsrat)

### Cultural Drivers

- not only economic explanations, also cultural / identity politics
- activated by economic shocks

@cantoniPersistenceActivationRightwing2019

- municipalities with nazi history 1933 = stonger support today
- AfD = party platofrm with low social stigma


### position & status

@kurerDisappointedExpectationsDownward2022 Status Discrodance

- calculated expected status based on parents occupation
- and difference to achieved status in reality
- especially men and eastern german high discordance

also @brianburgoonPositionalDeprivationSupport2019

- positional deprivation = income growth relative to other influences vote 
- slow = retreat from mainstream parties = feeling of losing out

### Literature on Minimum Wage and Politics

@zavodnyPoliticalEconomyMinimum2020a

-   political economy of minimum wages
-   focuses mostly on height of MW in different countries
    -   influencing factors on it

focus more general on labor market policies in general and support for populist vote @bergmanLabourMarketPolicies2022

- loss of status as driver


## Data

Vote Share AfD:
[Bundeswahlleiter](https://www.bundeswahlleiter.de/bundestagswahlen/2021/ergebnisse/bund-99.html)

-   calculated for the different Landkreise, not Wahlbezirke
-   Parliament Election on 09/2021

Minimum Wage: [Wirtschaft- und Sozialwissenschaftliches Institut (Hans
Böckler
Stiftung)](https://www.wsi.de/de/sonderauswertung-43368-mindestlohnanhebung-43374.htm)

based on:

-   BA Arbeit
-   Destatis
-   SOEP

=\> extrapolated by WSI

-   "sozialversicherungspflichtig beschäftigte" in district
-   part / full time, not MiniJob
-   projected for 01/10/2022

additional Data

-   Regionalstatistik des Bundes und der Länder
-   Regionatlas for Geodata

## Method

Linear Regressio with OLS Estimator

$$
\text{afd}_k = \beta_0 + \beta_1 \times \text{min}_k + \epsilon
$$

-   $\text{afd}_k$ AfD Vote Share in Landkreis $k$
-   $\text{min}_k$ Share of Minimum Wage Recipients in Landkreis $k$
-   $\beta_0$ Intercept
-   $\beta_1$ Slope
-   $\epsilon$ Error

## Results

basic Model

-   Estimated Coefficient = 0.7
-   statistically significant
-   R2 = 0.61

=\> 1% increase in MW recipoients =\> 0.7% increase in AfD share

Model with Eastern German Dummy

-   as visible on electoral Maps = eastern Germany more right wing vote
    share
-   dummy for ost (1 = eastern, 0 = not)
-   coefficient lower = 0.3
-   but still significant
-   R2 increase = 0.74

additional controls

-   not significant in all specificiations
-   do not change R2 much
-   no real impact
-   unemployment rate,
-   log GPD per Cap
-   log pop. density
-   foreigners rate
-   age

Test for Wahlergebnis of left poplist party (Die LINKE)

-   some way yes
-   but not really significant
-   as described by Piketty = left today not really "abgehängte"

## Discussion

Populism = multicausal, not only socioeconomic, but also
sociodemographic, cultural, ...


### my argument:

Minimum Wage is one of the channels translating abstract phenonema to populist vote:

phenomena  / factors (from literature):

- financial crisis
- austeritiy
- trade exposure

lead to: Minimum Wage / Low Wage sector

equals:
- lack of mobility
- status discordance
- general disappointment (inequality etc)

=> leads to populist voting

### factors: 

- financial crisis
    - shwon by @nauIncomePrecarityFinancial2019
    - crisis hits low income and middle income
    - creates income precarity
    - higher risk for insecuritry for all groups

- austerity
    - especially in euro-crisis @buschEuroCrisisAusterity
    - destruction of collective bargaining etc.
    - creates low wage sector 

- trade exposure
    - as shown by @utarWorkersFloodgatesLowWage2018
    - manufacturing workers => instable low wage service jobs

### channels

- lack of mobility
    - @dodinSocialMobilityGermany2021] adn
[@alesinaIntergenerationalMobilityPreferences2018]
    - if high share of MW = not many higher paying jobs
    - esp. eastern germans = not many high earning positions
    - future pessimism = easier for populists

- status discordancy @kurerDisappointedExpectationsDownward2022
    - MW = lowest status possible (except unemployed)
    - comparison to parent generation with higher status
        - either because east german = more middle class, compressed wage scale
        - pre financial crisis times = more job stability etc.

- positional deprivation @brianburgoonPositionalDeprivationSupport2019
    - rising share of top 10% @bartelsTopIncomesGermany2019
    - but not fast rising MW (except for 12 increase, but not much more)
    - lower end of wage scale (feels like down)
    - disappointed by system
    - see inequality
    - inequality =\> populism
    - ! in @brian, top income is good for left, while lower decie growth is good for right... 

- additional: instability of job, fear of losing joba nd being unemplyoed ling

### Controls

why minimum wage, not general low wage sector

- of course, MW kind of proxy for low wage sector
- but is additional psychological (very far end of wage scale)
- nobody below you (except unemployed)
- far away from high pay jobs of "winners" 


why vote for AfD, not Linke?

- shown by @berghExplainingRisePopulism2022, that unemployment = left, insecure job = right
- Linke = more for privileged voters (piketty)
- AfD more effective, much attention on them in public discourse


why now, not earlier?

- party changed attitude towards MW
    - earlier: jobkiller
    - now: important social policy (Grundsatzprogramm)
- social acceptance of Afd: Exit (Abstention) --> Voice
    - following @hirschmanExitVoiceLoyalty1972
- also: Inflation now hits very hard (2020f)
    - @toberHaushaltspezifischeTeuerungsraten2022
    - especially low wage in rural regions


endogeneity: Afd -> MW?

- not in power anywhere except Sonneberg and small districts
    - MW rise was before    
    - not long in power
- no change in economic structure there


labor market hypothesis?

- rejected (unemployment share not relevant)
- but may still have a place as irrational fear

Visualization of Theory:
```{mermaid}
%%{init: {'theme': "neutral"}}%%
graph LR

A([Financial Crisis]) & B([Austerity]) & C([Trade exposure]) --> MW((Mininum Wage))



MW --- D[Lack of mobility] & E[Status Discordance] & F[Positional Deprivation]
--> AfD((Populist voting))

```


### other channels:

eastern germany

-   hard hit by post-Wende turmoil, [@snowerCaringHandThat2006]
-   creation of Niedriglohnsektor due to TreuHand policy
-   cultural factors as well 
-   @mauLuettenKleinLeben2019 vast sociological study (shows difference)

inflation [@toberHaushaltspezifischeTeuerungsraten2022]

-   hits lower wage workers harder
-   especially rural regions with high share of car dependency


## Narrative Case Example: Sonneberg

## Conclusion

not stop from higher MW!

Policy

-   union bargaining incerase for higher wages
-   increase social mobility + decrease east west inequality

### future Research

-   characteristics of MW jobs
-   panel data set with more than one point in time

## To Do

- MW earners still rely on government support! (e.g Wohngeld)
    => nicht weiter ausgeführt
- [ ] controls for Bundesland FE
    => auch nicht
- Absatz zu Arbeitslosenrate und Zusammenhang
- l.116 Reformulate last sentence
    > würd ich etwas vorsichtiger ausdrücken, à la: eine kausale Interpretation (“influence”) ist möglich, falls wir alle relevanten controls inkludiert haben, was wir hoffen, aber nicht wissen. Außerdem guckst du dir nicht wirklich den influence von mw policies an, sondern vom Anteil an mw Empfängern, das ein Unterschied. 
- [x] MW wage **rate** umbennen in share ! (auch in Tabelle)
    => done
- l183 weiter ausformuleiren (Model 2)
    - > Das könnte man etwas genauer ausdrücken und vielleicht noch knowledge flexen: The coefficient on the share of minimum wage recipients substantially drops. This suggests that there is a positive correlation between a district being in East Germany and its minimum wage share. Model 3-1 thus suffered from a strong omitted variable bias: The positive coefficient on the “East” dummy and the positive correlation between “East Germany” and “minimum wage share” bias the coefficient on “minimum wage share” upwards in model 3-1.
- LINKE results in Appendix
    => done
- Ende Politik weniger
    - > Persönliche Präferenz, aber ich würde solche normativen calls etwas vorsichtiger ausdrücken: If the policy maker is interested in xy, he should consider doing z.
    => done, but could be better

Endogeineity Concerns adressen:

> Die ganze Discussion suggeriert etwas, was auch noch ausgeprochen werden muss: mw share hat vielleicht keinen kausalen Einfluss, sondern sowohl AfD als auch mw share sind Folgen eines third drivers, und dieser Driver kann sein: trade exposure, lack of upward mobility etc. Das acknowledgen! à la: Möglicherweise haben wir hier einen third driver (elaborieren), but to explore this is beyond the scope of this paper. Kandidaten für diesen third driver sind: GFC, Austerity, Trade, Lack of mobility, positional deprivation. 

> Ich sehe eigentlich den einzigen kausalen Effekt, der wirklich von mw share zu populist voting läuft in status discordance oder Stigmatisierung. Bei den anderen musst du dir nochmal überlegen, ob sowas wie lack of mobility wirklich eine Folge von mw share ist (wie im Bild suggeriert).

> Ok du schließt reverse causality aus aber es gibt noch andere Arten von Endogenität und die müssen dir hier Sorgen machen: Nämlich das, was auf deinem Bild die “other channels sind”. Wenn Ich Y auf X regresse, aber Y und X beide von Z getrieben werden, dann ist mein estimate bei der Y~X Regression falsch. Hier ist Z zb die GFC oder Austerity. Also lieber sagen: Reverse causality ist ausgeschlossen, aber endogeneity kann ein Problem sein.
